backend/requirements.txt
#Bot de Telegram
python-telegram-bot apscheduler
# API y Servidor
fastapi
uvicorn[standard]

# Base de Datos
sqlalchemy
alembic
psycopg2-binary

# Utilidades
python-dotenv
pydantic

async-generator==1.10
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
lxml==6.0.2
mypy_extensions==1.1.0
numpy==2.4.2
outcome==1.3.0.post0
packaging==26.0
pandas==3.0.0
pip==24.0
pycparser==3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
selenium==4.40.0
setuptools==65.5.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
zipp==3.23.0
